veersion 1.0 ready

+++ servlets/src/main/java/servlets/controllers/DeletePost.java
@@ -1,5 +1,6 @@
 package servlets.controllers;
 
+import core.Entities.Page;
 import jdbc.DaoProvider;
 import jdbc.dao.core.PostDao;
 import servlets.listeners.Initer;
@@ -13,6 +14,8 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
+import static servlets.listeners.DefaultSessionParams.AUTH;
+
 @WebServlet("/removepost")
 public class DeletePost extends HttpServlet {
     private PostDao postDao;
@@ -22,7 +25,9 @@ public class DeletePost extends HttpServlet {
         int page = Integer.parseInt(req.getParameter("id"));
         int post = Integer.parseInt(req.getParameter("post"));
 
-        postDao.delete(page, post);
+        Page me = (Page) req.getSession().getAttribute(AUTH);
+        if (me.getId() == page)
+            postDao.delete(page, post);
         resp.sendRedirect("/page");
     }
